Make commit for Rock-Paper-Scissors JS project
At this stage, the page uses the JS prompt and console to play a simple
rock-paper-scissors game. Using the prompt window, the program takes
user input for a player selection, validates the input, randomly generates
a computer selection, and determines a winner(or draw). The game() function
runs the simulation 5 times, keeping track of score (disregarding draws).
Currently, the program only displays results in the console, and requires
user input for every round of the game up front before displaying any
results.